tz1VZ13GQs5wo2eowSFP22WPEuvrDuCzZSnP
Having said that, it appears that there is indeed a network of addresses appearing to engage in market manipulation, particularly around Chick Population Tez. 

tz1VZ13GQs5wo2eowSFP22WPEuvrDuCzZSnP is the root address and receives funds from: 

1/ tz1UcF2dD8JrwKKAdea1eZWV3DRA6cXTu7jP --> Chick Population Tez

2/ tz1iSaNwPQaYsHNaLZZVR6zfEWhpzvcYkwDA --> Crypto-Man (this account is already under review by objkt.com)

3/ tz1Wb59foVCPVKdeLJf1mTR8evRXRCh3jnfv --> Funds Crypto-Man and another address tz1QYXrGMSS89chgY8Sd8pXeuDDJ5cXTUyZZ (see 4/)

4/ tz1QYXrGMSS89chgY8Sd8pXeuDDJ5cXTUyZZ --> Purchases Chick Population Tez as well as identified inter-account tx

5/ tz1dWTmPQbzqXxWsDspMNZRCncTXtjkbFe18 --> Purchases Chick Population Tez

6/ tz1gsod5XzKL9BYYfYjhKLVmuzdqQxUXyn3P --> Could be a Binance hot wallet that funds Chick Population Tez

7/ tz1dWTmPQbzqXxWsDspMNZRCncTXtjkbFe18 --> Purchases Chick Population Tez

Based on the above, all the addresses have been restricted, even if there appears to be a current focus on objkt.com, they are active on H=N and Rarible also.