tz1LFWn3KRW7KmX1pPhPQ4x7r2BBMPZScEhb
The underlying artworks are public domain, but the digital images are being made available via an open license with conditions - namely attribution, and use of the images must not use the Getty name (ie SUBJKT). None of the objkts I looked at are correctly following the license.
eg https://hicetnunc.art/objkt/668321
----
https://www.getty.edu/about/whatwedo/opencontent.html
Attribution to the Getty
Please use the following source credit when reproducing an Open Content image:
"Digital image courtesy of the Getty's Open Content Program."

Detailed caption information for individual artworks is provided as part of each download. When using open content images, you should not suggest or imply that the Getty endorses, approves of, or participated in your projects. The Getty name should not be used in the title or the name of your product, and it should not be used as a metadata search term, website name, or web address, or be large or prominent, as these uses and presentations tend to cause confusion among consumers by leading people to believe that the Getty Trust and its operating programs are affiliated with your project or service.