tz1THhVXbBf4zouuLmTG4GvpVc3vSMW59eEb
Although this looks like it was lifted from artvee which marks it as PD, that site does not properly provenance its scans & there are many works on it which breach terms & are not allowed to be used commercially. 

Unless you have your own copy of The book of baby birds / pictures in colour by E. J. Detmold published in 1912, you're not going to be able to use the original scans from https://babel.hathitrust.org/cgi/pt?id=njp.32101076149168&view=1up&seq=77&q1=rhea commercially, as these are flagged as https://www.hathitrust.org/access_use#pd-google Public Domain or Public Domain in the United States, Google-digitized: [...] Google requests that the images and OCR not be re-hosted, redistributed or used commercially. The images are provided for educational, scholarly, non-commercial purposes.